backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
python-dotenv==1.0.0
emails==0.6
jinja2==3.1.2
asyncpg==0.29.0
httpx==0.25.2
tenacity==8.2.3
